requirements.txt
streamlit
spacy
pdfminer.six
plotly